Rendered grand alliances are taken dynamically instead of hard coded.
The GrandAlliances::Index view now renders the list of grand_alliances passed up from the action instead of rendering a hard coded list.

Configuration has been added to seed the database with grand alliances at the start of a suite run and to clean the database of all tables except grand alliances.